fix(ui): simplify no data message
