+++ ui/src/style/components/tables.scss
@@ -113,6 +113,11 @@ tr.table-empty-state,
     }
   }
 }
+p.empty {
+  font-weight: 400;
+  font-size: 18px;
+  color: $g9-mountain;
+}
 
 /*
     Table Tabs
